chore: rename to tagsOverlay
