+++ libs/modelviewer/src/components/tags-overlay/TagsOverlay.tsx
